Increase compatibility

+++ webmanager/utils.py
@@ -15,13 +15,12 @@ class DataReader:
         output = {}
         for existing in FileManager.list_directory(f"cache/{cache_location}", ends_with=".json"):
             t_path = f"cache/{cache_location}/{existing}"
-            with FileManager.open_file(t_path) as file:
-                try:
-                    output[existing.replace('.json', '')] = json.load(file)
-                except Exception as e:
-                    print("Cache read error for %s: %s. Removing broken entry" % (t_path, str(e)))
-                    file.close()
-                    FileManager.remove_file(t_path)
+            data = FileManager.load_json_file(t_path)
+            if not data:
+                print("Cache read error for %s. Removing broken entry" % t_path)
+                FileManager.remove_file(t_path)
+
+            output[existing.replace('.json', '')] = data
 
         return output
 
@@ -89,10 +88,13 @@ class BuildingTemplateManager:
     def template_cache_list():
         output = {}
         for existing in FileManager.list_directory("templates/builder", ends_with=".txt"):
-            with FileManager.open_file(f"templates/builder/{existing}") as file:
-                output[existing] = BuildingTemplateManager.template_to_dict(
-                    [x.strip() for x in file.readlines()]
-                )
+            lines = FileManager.read_lines(f"templates/builder/{existing}")
+            if not lines:
+                continue
+
+            output[existing] = BuildingTemplateManager.template_to_dict(
+                [x.strip() for x in lines]
+            )
         return output
 
     @staticmethod
